Add possibility to insert header between the actions #46
